requirements.txt
# Core AI & HTTP
openai>=1.3.9
httpx>=0.24.1

# Web API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# File handling & env vars
python-multipart>=0.0.7
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Testing
pytest>=8.0.0

# Typing support
typing-extensions>=4.5.0
